+++ sdmx-dl-cli/src/main/java/sdmxdl/cli/ListSourcesCommand.java
@@ -61,11 +61,16 @@ public final class ListSourcesCommand implements Callable<Void> {
                 .columnOf("Website", SdmxWebSource::getWebsite, Formatter.onURL())
                 .columnOf("Monitor", SdmxWebSource::getMonitor, Formatter.onURI())
                 .columnOf("MonitorWebsite", SdmxWebSource::getMonitorWebsite, Formatter.onURL())
+                .columnOf("Languages", this::getLanguages, CsvUtil.fromIterable(Formatter.onString(), ','))
                 .build();
     }
 
     private String getDescription(SdmxWebSource source) {
         return source.getDescription(web.getLangs());
+    }
+
+    private Iterable<String> getLanguages(SdmxWebSource source) {
+        return () -> source.getDescriptions().keySet().stream().filter(lang -> !SdmxWebSource.ROOT_LANGUAGE.equals(lang)).iterator();
     }
 
     private Stream<SdmxWebSource> getRows() throws IOException {